More refactoring and polishing, and a couple of new features
Added a TiffWriter class so that opposite-endian TIFF files can be appended to correctly. Added the ability to append a single page of a TiffDocument to an existing file. Added a static method to TiffDocument to count the pages of a TIFF file without reading the whole file into memory.